# Upgrade to Docusaurus v3.2