--- a/docs/shadow-props.md
+++ b/docs/shadow-props.md
@@ -208,6 +208,8 @@ export default App;
 </TabItem>
 </Tabs>
 
+---
+
 # Reference
 
 ## Props