--- a/website/package.json
+++ b/website/package.json
@@ -48,10 +48,10 @@
     ]
   },
   "dependencies": {
-    "@docusaurus/core": "3.0.0-beta.0",
-    "@docusaurus/plugin-google-gtag": "3.0.0-beta.0",
-    "@docusaurus/plugin-pwa": "3.0.0-beta.0",
-    "@docusaurus/preset-classic": "3.0.0-beta.0",
+    "@docusaurus/core": "0.0.0-5878",
+    "@docusaurus/plugin-google-gtag": "0.0.0-5878",
+    "@docusaurus/plugin-pwa": "0.0.0-5878",
+    "@docusaurus/preset-classic": "0.0.0-5878",
     "docusaurus-plugin-sass": "^0.2.5",
     "esbuild-loader": "^2.21.0",
     "react": "^18.2.0",
@@ -60,7 +60,7 @@
     "sass": "^1.68.0"
   },
   "devDependencies": {
-    "@docusaurus/types": "3.0.0-beta.0",
+    "@docusaurus/types": "0.0.0-5878",
     "@react-native-website/lint-examples": "0.0.0",
     "@react-native-website/update-redirects": "0.0.0",
     "alex": "^10.0.0",