--- a/website/src/css/customTheme.scss
+++ b/website/src/css/customTheme.scss
@@ -625,7 +625,7 @@ a[class*="tagRegular"] {
     }
   }
 
-  .navbar__items div[class^="searchBox"] {
+  .navbar__items div[class^="navbarSearchContainer"] {
     padding-left: 14px;
   }