--- a/website/versioned_docs/version-0.70/shadow-props.md
+++ b/website/versioned_docs/version-0.70/shadow-props.md
@@ -98,6 +98,8 @@ const styles = StyleSheet.create({
 export default App;
 ```
 
+---
+
 # Reference
 
 ## Props