--- a/website/versioned_docs/version-0.71/shadow-props.md
+++ b/website/versioned_docs/version-0.71/shadow-props.md
@@ -209,6 +209,8 @@ export default App;
 </TabItem>
 </Tabs>
 
+---
+
 # Reference
 
 ## Props