--- a/website/versioned_docs/version-0.73/profile-hermes.md
+++ b/website/versioned_docs/version-0.73/profile-hermes.md
@@ -8,7 +8,7 @@ You can visualize JavaScript's performance in a React Native app using [Hermes](
 In this section, you will learn how to profile your React Native app running on Hermes and how to visualize the profile using [the Performance tab on Chrome DevTools](https://developers.google.com/web/tools/chrome-devtools/evaluate-performance/reference)
 
 :::caution
-Be sure to [enable Hermes in your app](Hermes) before you get started!
+Be sure to [enable Hermes in your app](hermes) before you get started!
 :::
 
 Follow the instructions below to get started profiling: