--- a/website/versioned_docs/version-0.73/shadow-props.md
+++ b/website/versioned_docs/version-0.73/shadow-props.md
@@ -208,6 +208,8 @@ export default App;
 </TabItem>
 </Tabs>
 
+---
+
 # Reference
 
 ## Props