@slorber 2024-03-29
@Simek v3.2 is released on npm and I updated the deps. This PR can be reviewed now ;)